initial setup for development data page, components reorganize into folders
