+++ packages/wheelhouse-frontend/src/components/reusables/Header.js
@@ -1,0 +1,9 @@
+import React from "react"; 
+
+const Header = () => {
+  return (
+    <h1 className="app-header">wheelhouse</h1>
+  );
+}; 
+
+export default Header;
