+++ packages/wheelhouse-frontend/src/components/reusables/Sidebar.js
@@ -1,14 +1,14 @@
 import React, { Component } from "react";
 import { Link } from "react-router-dom"; 
+import "../../Sidebar.css"; 
 
 class Sidebar extends Component {
   render() {
     return (
       <div className="sidebar-container">
         <ul>
+          <li><Link to="/development">Development</Link></li>
           <li><Link to="/pods">Pods</Link></li>
-          <li><Link to="/about">About</Link></li>
-          <li><Link to="/topics">Topics</Link></li>
         </ul>
       </div>
     );
