+++ packages/wheelhouse-frontend/src/components/reusables/Table.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../Table.css";
+import "../../Table.css";
 
 /*eslint-disable react/prop-types*/
 const Table = (props) => {
@@ -10,7 +10,7 @@ const Table = (props) => {
       </thead>
       <tbody>
         <tr> { props.descriptions } </tr>
-      </tbody>
+      </tbody>p
     </table>
   );
 };
