+++ packages/wheelhouse-frontend/src/index.css
@@ -3,3 +3,12 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
+
+.app-header {
+  text-align: center;
+  margin: 20px; 
+}
+
+.container {
+  max-width: 130rem;
+}
